Implemented Debt Prompts + Debt Resolution
+ Added contract breach penalty multiplier and contract debt collection toggle
+ Prompt window will come up on contract breach, and let players open the debt resolution screen to sell off trucks
+ Debt resolution window can also be opened from the office view

+++ unity-client/Assets/Scripts/DTO/ContractDTO.cs
@@ -9,7 +9,7 @@ public class ContractDTO : CompanyEntityDTO
     public ResourceType ResourceType { get; set; }
     public int DeliveredAmount { get; set; }
     public int TotalAmount { get; set; }
-    public int Payment { get; set; }
+    public float Payment { get; set; }
     public int ExpectedTick { get; set; }
     public int? DeliveredTick { get; set; }
     public int? AcceptedAtTick { get; set; }
